ClimbHistory: change "count" to a array of short integers
Save a few kilobytes of memory.  We don't need 2^63 samples.

+++ src/NMEA/ClimbHistory.hpp
@@ -39,7 +39,7 @@ class ClimbHistory {
   fixed vario[SIZE];
 
   /** Number of samples in each episode */
-  long count[SIZE];
+  unsigned short count[SIZE];
 
 public:
   void Clear();
